Address /query endpoint compatibility
This patch add some compatibily with the /query endpoint from
OLS-service.

For the request itself, everything should be identical now with the
exception of the "media_type" field. The "media_type" field has not been
included since the current implmentantion does not yet support streaming
responses.

For the response to a query, there are a lot of fields missing. These
also should be added as we start adding more features to Lightspeed
Core. At the moment, only the basic response is returned.

TODOs were left in the code pointing to those gaps.

Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>