Add an option to disable query system prompt
Since LCORE-293 (e60d7df1beca37a5909e5b066e39162c8fab889b), users can
set specify the system prompt in the config, but users can still
override it with their queries. This may not be desired in some cases,
so this commit adds an option to disable the query system prompt and
always use the configured system prompt (or the default one if none is
set). If a user tries to use the query system prompt when this option is
enabled, an HTTP 422 error is returned with a message explaining that
the query system prompt is not supported in this instance.