Containerfile COPY adjustment
We need to do some ugly hack in our assisted-chat repo to make our
pipelines work correctly, and it would make it easier for us if we could
configure where this Containerfile copies its files from.

It's a bit hard to explain the motivation, but if you don't mind it we
would appreciate it if we can have this until we can get rid of the hack
in the assisted-chat repo, when LSC has proper releases that we can use
out of the box.