Make MCP configurations unit-test more specific
Every time a configuration is added the
test_dump_configuration_with_more_mcp_servers() and
test_dump_configuration_with_one_mcp_server() are affected, it shouldn't
be the case since these tests should be focused on the MCP configuration.

For a general configuration test we already have the
test_dump_configuration().

This patch changes the MCP tests to be more focused on what they are
really asserting instead of a general configuration test that is already
covered as part of test_dump_configuration().

Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>