Support configuration env variable replacement
We want to support environment variable replacement in the configuration
similarly to how it is done in llama-stack.

Example use case / motivation:

The config, which since 3f7ed75e0e46766301dfbf622de3bfd0debb0736 can
contain a password to a database, requires users to move their entire
config to a Kubernetes secret, as it now contains sensitive information,
rather than a configmap. If instead we supported environment variable
replacement, users could keep their config in a configmap and just set
the password in a secret as k8s allows you to load secret values as
environment variables in pods.

Since users of lightspeed-stack are already familiar with the syntax
for environment variable replacement from llama-stack, we can use the
same function from llama-stack to handle this.