Allow authorization header to be omitted
Endpoints configured for the anybody (`*`) authorization should work
without an Authorization header.

This change allows requests without the header to be processed,
returning a default user identity, which will obviously receive the
anybody role (`*`) (because everyone receives it), so if the
configuration allows particular endpoints to be reachable by the `*`
role, guests (no auth header) users will be authorized to use it.