fix(auth): Use token in `AuthTuple`
While implementing authz I've made the authentication module put the
claims in the `AuthTuple` instead of putting the original token because
I assumed the original token was unnecessary and I didn't want to make
the authz module re-parse the token to get the claims.

That was a mistake, because we need to pass on the user token to the MCP
server for it to validate the user's access. So this commit changes the
`AuthTuple` to contain the original token instead of the claims.

The authz module has been updated to parse the claims from the token
instead of receiving them as a JSON string. Somewhat hackily, since we
don't want to re-verify the token signature, so we assume that the token
has already been verified during authentication and just decode the
claims from the middle section of the JWT token.