Regex operator for JWT role extraction
When we originally added the JWT role extraction operators, we made some
basic ones which thought would be enough, but we found that there's too
much complexity and variance in the various fields we find in Red Hat
SSO tokens, so we realized we're going to need a regex operator to
handle more complex matching scenarios

This commit simply adds a new `MATCH` operator to the `JsonPathOperator`
enum, and implements the logic to handle it in the `JwtRolesResolver`.

`_evaluate_operator` had to be modified to accept the entire
`JwtRoleRule` so that it can access the pre-compiled regex pattern for
better performance.