LCORE-395: config endpoint should not leak secrets
Right now in assisted-chat we're blocking the config endpoint
completely, even for employees, because it leaks secrets like:

- database password
- llama stack api key
- tls key password

This change makes sure those secrets are not included in the serialized
configuration. This way we can open up the config endpoint to employees
again (still not the public, because it contains things like system
prompt, which we would like to keep private).